Updated vimrc

+++ tools/sym_links.sh
@@ -2,17 +2,24 @@
 # $1 assumed to be pull location of dotfiles repo.
 CONFIG=$HOME/.config
 
+# Ensure dotfiles repo is defined
+if [ -z "$1" ]
+then
+      echo "\$1 must be dotfiles directory"
+      exit 1
+fi
+
 # symbolic link:
 #  ln -s target-path link-path
 
 # Alacritty
-mkdir $CONFIG/alacritty
-rm $CONFIG/alacritty/alacritty.yml
+mkdir -p $CONFIG/alacritty
+rm -f $CONFIG/alacritty/alacritty.yml
 echo "Linking $CONFIG/alacritty/alacritty.yml to $1/alacritty/alacritty.yml"
 ln -s $1/alacritty/alacritty.yml $CONFIG/alacritty/alacritty.yml
 
 # VS Code
-mkdir $CONFIG/Code/User
+mkdir -p $CONFIG/Code/User
 rm $CONFIG/Code/User/keybindings.json
 echo "Linking $CONFIG/Code/User/keybindings.json to $1/Code/User/keybindings.json"
 ln -s $1/Code/User/keybindings.json $CONFIG/Code/User/keybindings.json
@@ -21,7 +28,7 @@ echo "Linking $CONFIG/Code/User/settings.json to $1/Code/User/settings.json"
 ln -s $1/Code/User/settings.json $CONFIG/Code/User/settings.json
 
 # Tmux
-mkdir $CONFIG/tmux
+mkdir -p $CONFIG/tmux
 rm $CONFIG/tmux/tmux.conf
 echo "Linking $CONFIG/tmux/tmux.conf to $1/tmux/tmux.conf"
 ln -s $1/tmux/tmux.conf $CONFIG/tmux/tmux.conf
@@ -37,7 +44,7 @@ echo "Linking /etc/docker/deamon.json to $1/docker/daemon.json"
 sudo ln -s $1/docker/daemon.json /etc/docker/daemon.json
 
 # Tmux
-mkdir $CONFIG/tmux
+mkdir -p $CONFIG/tmux
 rm $CONFIG/tmux/tmux.conf
 echo "Linking $HOME/.ghc/ghci.conf to $1/haskell/ghci.conf"
 ln -s $1/haskell/ghci.conf $HOME/.ghc/ghci.conf
@@ -45,13 +52,13 @@ chmod go-w $HOME/.ghc
 chmod go-w $1/haskell/ghci.conf
 
 # Proxychains
-mkdir $HOME/.proxychains/
+mkdir -p $HOME/.proxychains/
 rm $HOME/.proxychains/proxychains.conf
 echo "Linking $HOME/.proxychains/proxychains.conf to $1/proxychains/proxychains.conf"
 ln -s $1/proxychains/proxychains.conf $HOME/.proxychains/proxychains.conf
 
 # PulseAudio
-mkdir $HOME/.config/pulse/
+mkdir -p $HOME/.config/pulse/
 rm $HOME/.config/pulse/default.pa
 echo "Linking $HOME/.config/pulse/default.pa to $1/pulse/default.pa"
 ln -s $1/pulse/default.pa $HOME/.config/pulse/default.pa
@@ -60,3 +67,12 @@ ln -s $1/pulse/default.pa $HOME/.config/pulse/default.pa
 rm $HOME/.gitconfig
 echo "Linking $HOME/.gitconfig to $1/git/.gitconfig"
 ln -s $1/git/.gitconfig $HOME/.gitconfig
+
+# Vim
+rm $HOME/.vimrc
+echo "Linking $HOME/.vimrc to $1/vim/.vimrc"
+ln -s $1/vim/.vimrc $HOME/.vimrc
+mkdir -p $HOME/.vim
+rm -rf $HOME/.vim/colors
+echo "Linking $HOME/.vim/colors to $1/vim/colors"
+ln -s $1/vim/colors $HOME/.vim/colors
